pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ac-server-gui"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.0",
    "pytest-mock>=3.10",
    "ruff>=0.1",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ac_server_gui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: tests that launch the actual AC server"]
addopts = "-m 'not integration'"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
strict = true
mypy_path = "src"
packages = ["ac_server_gui"]
